Move ErrorWriter and ErrorBase into include from lib.
